Inline Util.HTTP2.*
