+++ grapesy.cabal
@@ -75,8 +75,6 @@ library
       -Wprepositive-qualified-module
   exposed-modules:
       Network.GRPC.Util.Concurrency
-      Network.GRPC.Util.HTTP2
-      Network.GRPC.Util.HTTP2.Stream
       Network.GRPC.Util.Parser
       Network.GRPC.Util.PrettyVal
       Network.GRPC.Util.Thread
@@ -84,15 +82,10 @@ library
   hs-source-dirs:
       src
   build-depends:
-    , async         >= 2.2  && < 2.3
-    , binary        >= 0.8  && < 0.9
-    , bytestring    >= 0.10 && < 0.12
-    , http-types    >= 0.12 && < 0.13
-    , pretty-show   >= 1.10 && < 1.11
-    , stm           >= 2.5  && < 2.6
-
-    -- We need an unreleased version of http2 right now
-    , http2
+    , async       >= 2.2  && < 2.3
+    , bytestring  >= 0.10 && < 0.12
+    , pretty-show >= 1.10 && < 1.11
+    , stm         >= 2.5  && < 2.6
 
 test-suite test-grapesy
   import:
@@ -113,6 +106,7 @@ test-suite test-grapesy
     , grapesy
   build-depends:
       -- External dependencies
+    , binary
     , bytestring
     , exceptions
     , http-types
