--- src/Network/GRPC/Util/HTTP2.hs
@@ -1,16 +1,0 @@
-module Network.GRPC.Util.HTTP2 (
-    -- * General auxiliary
-    fromHeaderTable
-  ) where
-
-import Data.Bifunctor
-import Network.HPACK qualified as HPACK
-import Network.HPACK.Token qualified as HPACK
-import Network.HTTP.Types qualified as HTTP
-
-{-------------------------------------------------------------------------------
-  General auxiliary
--------------------------------------------------------------------------------}
-
-fromHeaderTable :: HPACK.HeaderTable -> [HTTP.Header]
-fromHeaderTable = map (first HPACK.tokenKey) . fst
